MGMT-1883 - Add update cluster install config api (#230)
* MGMT-1929 - Add required fields to install config

This adds the missing fields to the install config struct.
They will later be made customizable.

* Add swagger endpoint for update cluster install config and regenerate files

* Add clusters column for install config overrides

This will be a json string that can later be parsed and merged
into our default install config

It will be set by the previously added UpdateClusterInstallConfig
endpoint.

* Save the install config overrides to the cluster

* Apply install config overrides

This commit applies the cluster config override json over the
previous install config struct to apply the user values.

* Add an example of the install config override payload to swagger.yaml

* Move install config validation logic to the installcfg package