MGMT-2320: fixed timezone comparision in event_test.go
When creating time.Time() object to store in database it initialized
with Local timezone. But when loading events from database the returned
time objects are converted to UTC timezone.
In this commit the time retrieved from the event list is specified to
be Local and by that we can be more accurate when comparing between times.