NO-ISSUE: Fixed get_service_host/get_service_port for local mode
Since the beginning get_service_host/get_service_port functions used
minikube command with invalid '-n' flag, which applied to kubectl.
That means that 'get_service_url' function never worked in local mode,
or if '--target' was not provided.
In such cases we should use the existing functionality that for some
reason was not used in minikube mode.